fix pacman sed
sed expressions using the a, c or i commands cant be separated with ; and require a newline or -e to do so instead. since the ILoveCandy is the only expression in the sequence that uses one of those commands i just moved it to the end which seemed to fix the problem. also shortened the parallel downloads sed. chaining expressions is addressed in section 3.8.1 of the gnu sed manual